refactor(prs): round 1 — deduplicate interfaces + extract timeAgo utility
Round-1 cloud findings:

MUST FIX (applied):
- [Gemini, PrTicker.tsx + usePrCache.ts] PrRow and RepoSummary interfaces
  duplicated — exported from usePrCache.ts, removed local copies in
  PrTicker.tsx, imported via `import type { PrRow, RepoSummary }`.
- [Gemini, PrTicker.tsx] timeAgo reimplemented locally — extracted to
  apps/web/src/lib/time.ts as shared utility, imported in PrTicker.tsx.

Per cloud-thrift policy, NOT cycling this PR — local swarm clean
is the signal, existing cloud review stands.
